Migrate to one single schema.ts file
