+++ drizzle.config.ts
@@ -2,7 +2,7 @@ import { env } from "@/env";
 import type { Config } from "drizzle-kit";
 
 export default {
-	schema: "./src/server/db/schema/index.ts",
+	schema: "./src/server/db/schema.ts",
 	out: "./drizzle",
 	dialect: "postgresql",
 	strict: true,
