+++ drizzle/0006_strange_fabian_cortez.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "Teensy" DROP CONSTRAINT "Teensy_ownerId_user_id_fk";
