+++ drizzle/meta/_journal.json
@@ -43,6 +43,13 @@
 			"when": 1725830457735,
 			"tag": "0005_demonic_xorn",
 			"breakpoints": true
+		},
+		{
+			"idx": 6,
+			"version": "7",
+			"when": 1725831440697,
+			"tag": "0006_strange_fabian_cortez",
+			"breakpoints": true
 		}
 	]
 }
