--- src/server/db/migrate.ts
@@ -1,18 +1,0 @@
-import { migrate } from "drizzle-orm/postgres-js/migrator";
-
-import { client, db } from "./index";
-
-async function main() {
-	// biome-ignore lint/suspicious/noConsoleLog: This is a dev script
-	console.log("Running migrations");
-	await migrate(db, { migrationsFolder: "drizzle" });
-	// biome-ignore lint/suspicious/noConsoleLog: This is a dev script
-	console.log("Migrations finished");
-	await client.end();
-}
-
-void main().catch((e) => {
-	console.error("Error running migrations");
-	console.error(e);
-	process.exit(1);
-});
