--- src/server/db/schema/account.ts
@@ -1,34 +1,0 @@
-import { integer, pgTable, primaryKey, text } from "drizzle-orm/pg-core";
-import { createInsertSchema, createSelectSchema } from "drizzle-zod";
-import type { AdapterAccount } from "next-auth/adapters";
-import type { z } from "zod";
-import { user } from "./user";
-
-export const account = pgTable(
-	"account",
-	{
-		userId: text("userId")
-			.notNull()
-			.references(() => user.id, { onDelete: "cascade", onUpdate: "cascade" }),
-		type: text("type").$type<AdapterAccount["type"]>().notNull(),
-		provider: text("provider").notNull(),
-		providerAccountId: text("providerAccountId").notNull(),
-		refresh_token: text("refresh_token"),
-		access_token: text("access_token"),
-		expires_at: integer("expires_at"),
-		token_type: text("token_type"),
-		scope: text("scope"),
-		id_token: text("id_token"),
-		session_state: text("session_state"),
-	},
-	(account) => ({
-		compoundKey: primaryKey({
-			columns: [account.provider, account.providerAccountId],
-		}),
-	}),
-);
-
-export const selectAccountSchema = createSelectSchema(account);
-export const insertAccountSchema = createInsertSchema(account);
-
-export type Account = z.infer<typeof selectAccountSchema>;
