--- src/server/db/schema/index.ts
@@ -1,10 +1,0 @@
-export * from "./enum";
-export * from "./user";
-export * from "./verificationToken";
-export * from "./visit";
-export * from "./globalVisits";
-export * from "./account";
-export * from "./session";
-export * from "./teensy";
-export * from "./expiredTeensy";
-export * from "./relations";
